+++ angular/scripts/controllers/home/mainCtrl.js
@@ -153,4 +153,20 @@ angular.module('theOneBlog')
   // });
 
 }])
+  .controller('CateListCtrl', ['$scope', '$filter', 'articlesByCate', 'curpage', function ($scope, $filter, articlesByCate, curpage) {
+
+    $scope.pageNav = {
+      maxSize:5,
+      bigTotalItems:articlesByCate.cate.articleNum,
+      bigCurrentPage:curpage
+    };
+
+    $scope.articleList = $filter('ArticlesByCateWordLimit')(articlesByCate.articleList);
+    
+    $scope.cate = articlesByCate.cate;
+
+
+
+
+  }])
   ;