+++ app/controllers/article.js
@@ -5,7 +5,8 @@
 'use strict';
 
 var mongoose = require('mongoose'),
-  // async = require('async'),
+  async = require('async'),
+  Cate = mongoose.model('Cate'),
   Article = mongoose.model('Article');
   // _ = require('underscore'),
   // Tag = mongoose.model('Tag');
@@ -44,6 +45,39 @@ exports.getById = function (req, res, next) {
     res.json(article);
     res.end();
   });
+};
+
+// 分类文章
+exports.getListByCate = function (req, res, next) {
+  var cateAlias = req.params.alias,
+  curPage = req.params.page || 1,
+  pageCount = 12,
+  skipNum = (curPage - 1) * pageCount,
+  result = {};
+
+  async.waterfall([
+    function (cb) {
+      Cate.findByAlias(cateAlias, cb);
+    },function (targetCate, cb) {
+      result.cate = targetCate;
+
+      Article
+        .find()
+        .where({cate:targetCate._id})
+        .skip(skipNum)
+        .limit(pageCount)
+        .exec(cb);
+
+    }],function (err, articles) {
+      result.articleList = articles;
+
+      if(err) {
+        return next(err);
+      }
+      res.json(result);
+      res.end();
+
+    });
 };
 
 // 文章搜索