+++ app/routers/routerHome.js
@@ -19,7 +19,8 @@ module.exports = function (app) {
 
   // 文章列表
   app.get('/article/list', articleCtrl.list);
-  app.get('/article/id/:id',articleCtrl.getById);
+  app.get('/article/id/:id', articleCtrl.getById);
+  app.get('/article/cate/:alias/:page', articleCtrl.getListByCate);
 
   app.get('/search/:searchWord', searchCtrl.goAngthing);